+++ src/main/market_and_gear/LightningSpell.java
@@ -33,7 +33,7 @@ public class LightningSpell extends Spell {
      * @param mana Mana required to use this Spell.
      * @param damage amount of damage caused by this Spell.
      */
-    public LightningSpell(String name, int price, Level minLevel, Mana mana, int damage) {
+    public LightningSpell(String name, int price, Level minLevel, Mana mana, double damage) {
         super(name, price, minLevel, damage, mana, AbilityBuilder.baseAgilityAbility());
     }
